# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BomItemResponse": {
        "description": "BOM 行（B.8：material_code/quantity_per——quantity_per 即用量）。",
        "properties": {
          "material_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Code"
          },
          "quantity_per": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity Per"
          }
        },
        "title": "BomItemResponse",
        "type": "object"
      },
      "BomResponse": {
        "description": "GET /tools/bom 响应（product_code/bom_version/items[]）。",
        "properties": {
          "bom_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bom Version"
          },
          "evidence_hint": {
            "$ref": "#/components/schemas/EvidenceHint"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BomItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "product_code": {
            "title": "Product Code",
            "type": "string"
          }
        },
        "required": [
          "product_code",
          "items",
          "evidence_hint"
        ],
        "title": "BomResponse",
        "type": "object"
      },
...
      "CustomerResponse": {
        "description": "GET /tools/customers/{code} 响应（name/level/attributes）。",
        "properties": {
          "attributes": {
            "additionalProperties": true,
            "title": "Attributes",
            "type": "object"
          },
          "customer_code": {
            "title": "Customer Code",
            "type": "string"
          },
          "evidence_hint": {
            "$ref": "#/components/schemas/EvidenceHint"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "customer_code",
          "name",
          "attributes",
          "evidence_hint"
        ],
        "title": "CustomerResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "错误体（附录 B.0）；运行时 extra 字段按需并入。",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "统一错误响应包裹：{\"error\": {code, message, request_id}}。",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "EvidenceHint": {
        "description": "关联对象/事件（该对象最新 {TYPE}_SNAPSHOT 事件；无则 event_id=null）。",
        "properties": {
          "event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "object_id": {
            "format": "uuid",
            "title": "Object Id",
            "type": "string"
          }
        },
        "required": [
          "object_id"
        ],
        "title": "EvidenceHint",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InventoryResponse": {
        "description": "GET /tools/inventory 响应（warehouses[] + total_available + snapshot_at）。",
        "properties": {
          "evidence_hint": {
            "$ref": "#/components/schemas/EvidenceHint"
          },
          "material_code": {
            "title": "Material Code",
            "type": "string"
          },
          "material_id": {
            "format": "uuid",
            "title": "Material Id",
            "type": "string"
          },
          "snapshot_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot At"
          },
          "total_available": {
            "title": "Total Available",
            "type": "number"
          },
          "warehouses": {
            "items": {
              "$ref": "#/components/schemas/WarehouseItem"
            },
            "title": "Warehouses",
            "type": "array"
          }
        },
        "required": [
          "material_code",
          "material_id",
          "warehouses",
          "total_available",
          "evidence_hint"
        ],
        "title": "InventoryResponse",
        "type": "object"
      },
...
      "OrderDetail": {
        "description": "订单详情（B.8：摘要结构 + lines 行明细）。",
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "customer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "delivery_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Date"
          },
          "evidence_hint": {
            "$ref": "#/components/schemas/EvidenceHint"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/OrderLineItem"
            },
            "title": "Lines",
            "type": "array"
          },
          "object_id": {
            "format": "uuid",
            "title": "Object Id",
            "type": "string"
          },
          "order_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Date"
          },
          "order_no": {
            "title": "Order No",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "order_no",
          "object_id",
          "currency",
          "status",
          "evidence_hint",
          "lines"
        ],
        "title": "OrderDetail",
        "type": "object"
      },
...
      "OrderListResponse": {
        "description": "GET /tools/orders 响应（摘要列表；游标分页字段占位）。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "OrderListResponse",
        "type": "object"
      },
...
      "PurchaseOrderListResponse": {
        "description": "GET /tools/purchase-orders 响应（items + next_cursor）。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PurchaseOrderItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "PurchaseOrderListResponse",
        "type": "object"
      },
...
      "SupplierLeadTimesResponse": {
        "description": "GET /tools/supplier-lead-times 响应（lead_times[] + updated_at）。",
        "properties": {
          "evidence_hint": {
            "$ref": "#/components/schemas/EvidenceHint"
          },
          "lead_times": {
            "items": {
              "$ref": "#/components/schemas/LeadTimeItem"
            },
            "title": "Lead Times",
            "type": "array"
          },
          "supplier_code": {
            "title": "Supplier Code",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "supplier_code",
          "lead_times",
          "evidence_hint"
        ],
        "title": "SupplierLeadTimesResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WarehouseItem": {
        "description": "库存仓库粒度快照（B.8：warehouse/available/reserved）。",
        "properties": {
          "available": {
            "title": "Available",
            "type": "number"
          },
          "reserved": {
            "title": "Reserved",
            "type": "number"
          },
          "warehouse": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warehouse"
          }
        },
        "required": [
          "available",
          "reserved"
        ],
        "title": "WarehouseItem",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "EDP 数据平台开放 API：认证、主数据登记（registry）与事件批量入库（events）。",
    "title": "EDP Data Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tools/bom": {
      "get": {
        "description": "产品 ACTIVE BOM；产品/版本不存在 → 404。",
        "operationId": "get_bom_api_v1_tools_bom_get",
        "parameters": [
          {
            "description": "产品 code",
            "in": "query",
            "name": "product_code",
            "required": true,
            "schema": {
              "description": "产品 code",
              "title": "Product Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BomResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "UNAUTHENTICATED：未认证或凭据无效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "FORBIDDEN：权限不足（scope/权限不满足）；TENANT_SUSPENDED：租户已暂停或状态异常"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "NOT_FOUND：资源不存在（跨租户统一 404，不泄露存在性）"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "METHOD_NOT_ALLOWED：方法不允许"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "产品 BOM（ACTIVE 版本 + 用量行）",
        "tags": [
          "tools"
        ]
      }
    },
    "/api/v1/tools/customers/{customer_code}": {
      "get": {
        "description": "客户主数据；不存在/跨租户统一 404 NOT_FOUND。",
        "operationId": "get_customer_api_v1_tools_customers__customer_code__get",
        "parameters": [
          {
            "in": "path",
            "name": "customer_code",
            "required": true,
            "schema": {
              "title": "Customer Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "UNAUTHENTICATED：未认证或凭据无效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "FORBIDDEN：权限不足（scope/权限不满足）；TENANT_SUSPENDED：租户已暂停或状态异常"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "NOT_FOUND：资源不存在（跨租户统一 404，不泄露存在性）"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "METHOD_NOT_ALLOWED：方法不允许"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "客户主数据（name/level/attributes）",
        "tags": [
          "tools"
        ]
      }
    },
    "/api/v1/tools/inventory": {
      "get": {
        "description": "物料库存；物料不存在 → 404，无库存行 → 200 空 warehouses。",
        "operationId": "get_inventory_api_v1_tools_inventory_get",
        "parameters": [
          {
            "description": "物料 code",
            "in": "query",
            "name": "material_code",
            "required": true,
            "schema": {
              "description": "物料 code",
              "title": "Material Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "UNAUTHENTICATED：未认证或凭据无效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "FORBIDDEN：权限不足（scope/权限不满足）；TENANT_SUSPENDED：租户已暂停或状态异常"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "NOT_FOUND：资源不存在（跨租户统一 404，不泄露存在性）"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "METHOD_NOT_ALLOWED：方法不允许"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "物料库存（仓库粒度 + 合计可用量）",
        "tags": [
          "tools"
        ]
      }
    },
    "/api/v1/tools/orders": {
      "get": {
        "description": "订单摘要列表（B.8：同详情结构数组，不含 lines 明细）。",
        "operationId": "list_orders_api_v1_tools_orders_get",
        "parameters": [
          {
            "description": "客户 code 过滤",
            "in": "query",
            "name": "customer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "客户 code 过滤",
              "title": "Customer"
            }
          },
          {
            "description": "订单状态过滤",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "订单状态过滤",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "UNAUTHENTICATED：未认证或凭据无效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "FORBIDDEN：权限不足（scope/权限不满足）；TENANT_SUSPENDED：租户已暂停或状态异常"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "METHOD_NOT_ALLOWED：方法不允许"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "订单摘要列表（customer/status 过滤）",
        "tags": [
          "tools"
        ]
      }
    },
    "/api/v1/tools/orders/{order_no}": {
      "get": {
        "description": "订单详情；不存在/跨租户统一 404 NOT_FOUND（不泄露存在性）。",
        "operationId": "get_order_api_v1_tools_orders__order_no__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_no",
            "required": true,
            "schema": {
              "title": "Order No",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "UNAUTHENTICATED：未认证或凭据无效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "FORBIDDEN：权限不足（scope/权限不满足）；TENANT_SUSPENDED：租户已暂停或状态异常"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "NOT_FOUND：资源不存在（跨租户统一 404，不泄露存在性）"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "METHOD_NOT_ALLOWED：方法不允许"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "订单详情（含行明细与客户）",
        "tags": [
          "tools"
        ]
      }
    },
    "/api/v1/tools/purchase-orders": {
      "get": {
        "description": "采购单列表（B.8：items + next_cursor；游标分页字段本轮占位）。",
        "operationId": "list_purchase_orders_api_v1_tools_purchase_orders_get",
        "parameters": [
          {
            "description": "物料 code 过滤",
            "in": "query",
            "name": "material_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "物料 code 过滤",
              "title": "Material Code"
            }
          },
          {
            "description": "采购单状态过滤",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "采购单状态过滤",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOrderListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "UNAUTHENTICATED：未认证或凭据无效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "FORBIDDEN：权限不足（scope/权限不满足）；TENANT_SUSPENDED：租户已暂停或状态异常"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "METHOD_NOT_ALLOWED：方法不允许"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "采购单列表（material_code/status 过滤）",
        "tags": [
          "tools"
        ]
      }
    },
    "/api/v1/tools/supplier-lead-times": {
      "get": {
        "description": "供应商交期；供应商不存在 → 404，无交期行 → 200 空 lead_times。",
        "operationId": "list_supplier_lead_times_api_v1_tools_supplier_lead_times_get",
        "parameters": [
          {
            "description": "供应商 code",
            "in": "query",
            "name": "supplier_code",
            "required": true,
            "schema": {
              "description": "供应商 code",
              "title": "Supplier Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierLeadTimesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "UNAUTHENTICATED：未认证或凭据无效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "FORBIDDEN：权限不足（scope/权限不满足）；TENANT_SUSPENDED：租户已暂停或状态异常"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "NOT_FOUND：资源不存在（跨租户统一 404，不泄露存在性）"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "METHOD_NOT_ALLOWED：方法不允许"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "供应商交期（物料粒度）",
        "tags": [
          "tools"
        ]
      }
    },
...
  }
}